Added login requirement for restricted actions

+++ products/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from .models import Product, Category
 from .forms import AddProductForm
 
@@ -60,8 +61,13 @@ def all_products(request):
     return render(request, 'products/products.html', context)
 
 
+@login_required
 def add_product(request):
     """ View for adding products to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Oops, you need to be authorised to do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = AddProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -81,8 +87,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
-    """ View for editing a product in the store """
+    """ View for editing a product in the store """    
+    if not request.user.is_superuser:
+        messages.error(request, 'Oops, you need to be authorised to do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = AddProductForm(request.POST, request.FILES, instace=product)
@@ -105,8 +116,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ View for delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Oops, you need to be authorised to do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
